chore: Update JOBS/priority_queue API
